requirements.txt
aiohttp==3.5.4
Pillow==5.4.1
gunicorn==19.9.0
aiopg==0.16.0
SQLAlchemy==1.3.0
